+++ example/src/App.tsx
@@ -44,7 +44,7 @@ function App() {
         background: '#f3f3f3',
       }}
     >
-      <DataSheetGrid data={data} onChange={setData} columns={columns} />
+      <DataSheetGrid value={data} onChange={setData} columns={columns} />
     </div>
   )
 }