+++ website/src/demos/renderWhenScrolling.tsx
@@ -17,7 +17,7 @@ export default () => {
 
   return (
     <DataSheetGrid
-      data={data}
+      value={data}
       onChange={setData}
       columns={[
         {