+++ website/src/demos/underlyingData.tsx
@@ -21,7 +21,7 @@ export default () => {
   return (
     <>
       <DataSheetGrid
-        data={data}
+        value={data}
         onChange={setData}
         columns={[
           { ...keyColumn('active', checkboxColumn), title: 'Active' },