+++ website/src/demos/xxl.tsx
@@ -28,7 +28,7 @@ export default () => {
 
   return (
     <DataSheetGrid
-      data={data}
+      value={data}
       onChange={setData}
       columns={[
         { ...keyColumn('active', checkboxColumn), title: 'Active' },